move more CanBeExposed from factory_client_access
