+++ pkg/kubectl/cmd/util/factory_client_access.go
@@ -47,7 +47,6 @@ import (
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
-	"k8s.io/kubernetes/pkg/apis/apps"
 	api "k8s.io/kubernetes/pkg/apis/core"
 	"k8s.io/kubernetes/pkg/apis/extensions"
 	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
@@ -457,17 +456,6 @@ func Contains(resourcesList []*metav1.APIResourceList, resource schema.GroupVers
 
 func (f *ring0Factory) Generators(cmdName string) map[string]kubectl.Generator {
 	return DefaultGenerators(cmdName)
-}
-
-func (f *ring0Factory) CanBeExposed(kind schema.GroupKind) error {
-	switch kind {
-	case api.Kind("ReplicationController"), api.Kind("Service"), api.Kind("Pod"),
-		extensions.Kind("Deployment"), apps.Kind("Deployment"), extensions.Kind("ReplicaSet"), apps.Kind("ReplicaSet"):
-		// nothing to do here
-	default:
-		return fmt.Errorf("cannot expose a %s", kind)
-	}
-	return nil
 }
 
 // this method exists to help us find the points still relying on internal types.
