+++ pkg/kubectl/polymorphichelpers/canbeexposed_test.go
@@ -1,0 +1,50 @@
+/*
+Copyright 2018 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package polymorphichelpers
+
+import (
+	"testing"
+
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	api "k8s.io/kubernetes/pkg/apis/core"
+)
+
+func TestCanBeExposed(t *testing.T) {
+	tests := []struct {
+		kind      schema.GroupKind
+		expectErr bool
+	}{
+		{
+			kind:      api.Kind("ReplicationController"),
+			expectErr: false,
+		},
+		{
+			kind:      api.Kind("Node"),
+			expectErr: true,
+		},
+	}
+
+	for _, test := range tests {
+		err := canBeExposed(test.kind)
+		if test.expectErr && err == nil {
+			t.Error("unexpected non-error")
+		}
+		if !test.expectErr && err != nil {
+			t.Errorf("unexpected error: %v", err)
+		}
+	}
+}
